+++ NEWS.md
@@ -15,3 +15,7 @@ the number of dimensions to reach a target *pseudo R squared*
 - added ggplotly() functionality for output MDS plots
 - in `ggplotSampleMDS()`, added `flipXAxis`, `flipYAxis` 
 to possibly ease low dimensional projection comparisons
+- in `ggplotSampleMDS()`, added `dispdisplayArrowLabels` to discard
+the arrow labels in biplot, while keeping the ability to interactively show
+them when using ggplot object in `ggplotly()`
+